requirements.txt
# Flask Framework
Flask==3.0.0
Jinja2==3.1.2
Werkzeug==3.0.1

# Supabase Client (works with httpx 0.27.x)
supabase==2.8.1

# Google Generative AI (compatible with httpx 0.27.x)
google-generativeai==0.8.3

# HTTP Client (required by both supabase and google-genai)
httpx==0.27.2

# WebSockets (required by google-genai)
websockets==13.1

# Encryption & Security
cryptography==41.0.7
argon2-cffi==23.1.0

# Flask Extensions
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
WTForms==3.1.1

# Environment & Configuration
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.7

# WSGI Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code Quality
black==23.12.1
flake8==6.1.0
